+++ src/Products/CPUtils/Extensions/utils.py
@@ -138,7 +138,7 @@ def log_list(lst, line, logger=None, level='info'):
     lst.append(line)
 
 
-def object_link(obj, view='view', attribute='Title', content=''):
+def object_link(obj, view='view', attribute='Title', content='', target=''):
     """ Returns an html link for the given object """
     from Products.CMFPlone.utils import safe_unicode
     href = view and "%s/%s" % (obj.absolute_url(), view) or obj.absolute_url()
@@ -148,7 +148,9 @@ def object_link(obj, view='view', attribute='Title', content=''):
         content = getattr(obj, attribute)
         if callable(content):
             content = content()
-    return u'<a href="%s">%s</a>' % (href, safe_unicode(content))
+    if target:
+        target = ' target="{}"'.format(target)
+    return u'<a href="%s"%s>%s</a>' % (href, target, safe_unicode(content))
 
 ###############################################################################