+++ CHANGES.rst
@@ -7,7 +7,8 @@ Changelog
 
 - Fix for password validation that was validating the hash
   [bleybaert]
-
+- Added uid method to display current uid
+  [sgeulette]
 
 1.17 (2020-04-02)
 -----------------